refactor(cli): share local target activation guard
